examples/sockthief/requirements.txt
#longchain[slack]
python-dotenv